requirements.txt
tensorflow
streamlit
opencv-python
numpy
pillow
pandas
scikit-learn